Adds licensing information from the Github API
For every formula that lists a Github repository as its homepage, the
Github License API was attempted to be used to retrieve license information.
If the Github API returned a useful result, it has been used.

This commit also adds quite a few formulae that weren't in the original hash.

Signed-off-by: zachwick <zach@zachwick.com>